+++ terminal.c
@@ -618,6 +618,14 @@ term_set_fonts(struct terminal *term, struct fcft_font *fonts[static 4])
         term->fonts[i] = fonts[i];
     }
 
+    for (size_t i = 0; i < ALEN(term->box_drawing); i++) {
+        if (term->box_drawing[i] != NULL) {
+            pixman_image_unref(term->box_drawing[i]->pix);
+            free(term->box_drawing[i]);
+            term->box_drawing[i] = NULL;
+        }
+    }
+
     const int old_cell_width = term->cell_width;
     const int old_cell_height = term->cell_height;
 
@@ -1411,6 +1419,13 @@ term_destroy(struct terminal *term)
         fcft_destroy(term->fonts[i]);
     for (size_t i = 0; i < 4; i++)
         free(term->font_sizes[i]);
+
+    for (size_t i = 0; i < ALEN(term->box_drawing); i++) {
+        if (term->box_drawing[i] != NULL) {
+            pixman_image_unref(term->box_drawing[i]->pix);
+            free(term->box_drawing[i]);
+        }
+    }
 
     free(term->search.buf);